Move RPC tracking to SpatialMetrics
